+++ bin/faucet/src/api/get_tokens.rs
@@ -1,6 +1,6 @@
 use axum::Json;
 use axum::extract::{Query, State};
-use axum::http::StatusCode;
+use axum::http::{HeaderMap, HeaderValue, StatusCode};
 use axum::response::{IntoResponse, Response};
 use miden_client::account::{AccountId, AccountIdError};
 use miden_faucet_lib::requests::{MintError, MintRequest, MintRequestSender};
@@ -127,7 +127,7 @@ pub enum GetTokenError {
 impl GetTokenError {
     fn status_code(&self) -> StatusCode {
         match self {
-            Self::InvalidRequest(MintRequestError::PowError(PowError::RateLimited)) => {
+            Self::InvalidRequest(MintRequestError::PowError(PowError::RateLimited(_))) => {
                 StatusCode::TOO_MANY_REQUESTS
             },
             Self::InvalidRequest(_) | Self::MintError(_) => StatusCode::BAD_REQUEST,
@@ -164,12 +164,26 @@ impl GetTokenError {
             },
         }
     }
+
+    fn headers(&self) -> HeaderMap {
+        let mut headers = HeaderMap::new();
+        if let Self::InvalidRequest(MintRequestError::PowError(PowError::RateLimited(timestamp))) =
+            self
+        {
+            headers.insert(
+                axum::http::header::RETRY_AFTER,
+                // SAFETY: timestamp is always a valid string
+                HeaderValue::from_str(&timestamp.to_string()).unwrap(),
+            );
+        }
+        headers
+    }
 }
 
 impl IntoResponse for GetTokenError {
     fn into_response(self) -> Response {
         self.trace();
-        (self.status_code(), self.user_facing_error()).into_response()
+        (self.headers(), (self.status_code(), self.user_facing_error())).into_response()
     }
 }